refactored, tests working
