+++ .circleci/config.yml
@@ -1,14 +1,13 @@
-  
+# Use the latest 2.1 version of CircleCI pipeline process engine. See: https://circleci.com/docs/2.0/configuration-reference
 version: 2.1
+# Use a package of configuration called an orb.
 orbs:
-  node: circleci/node@4.0.1
-jobs:
-  build:
-    docker:
-      - image: circleci/node:12.18.3
-      - image: circleci/mongo:4.4.1-ram
-    steps:
-      - checkout
-      - run: npm install
-      - run: npm test
-      - run: npm db-seed
+  # Declare a dependency on the welcome-orb
+  welcome: circleci/welcome-orb@0.4.1
+# Orchestrate or schedule a set of jobs
+workflows:
+  # Name the workflow "welcome"
+  welcome:
+    # Run the welcome/run job in its own container
+    jobs:
+      - welcome/run
