+++ client/src/components/Question.jsx
@@ -7,13 +7,12 @@ import styled from 'styled-components';
 const Wrapper = styled.section`
 	margin: auto;
 	width: 50%;
-	border-bottom: 1px solid grey;
 	border-top: 1px solid grey;
 	padding: 10px;
 	position: relative;
 	font-family: Stuart, Georgia, serif;
-	padding-bottom: 50 px;
-	margin-bottom: 50px;
+	padding-bottom: 10 px;
+	margin-bottom: 10px;
 `;
 
 const HelpfulButton = styled.button`
@@ -36,11 +35,17 @@ const AnswerButton = styled.button`
 	border-radius: 2px;
 `;
 
+const User = styled.div`
+	font-family: "Roboto","Helvetica Neue","Helvetica","Arial",sans-serif;
+	font-size: 16px;
+	font-weight: 700;
+`;
+
 const Question = (props) => (
 	<Wrapper>
 	<div>
 		<AnswerNumber><p>{props.question.answers.length}</p><p>{'\n'}answers</p></AnswerNumber>
-		<div>{props.question.user} <span>&#183;</span> {moment(props.question.created_at).startOf('hour').fromNow()}</div>
+		<div><User>{props.question.user}</User><span>&#183;</span> {moment(props.question.created_at).startOf('hour').fromNow()}</div>
 		<h3>{props.question.question_body}</h3>
 		<AnswerButton>Answer the question</AnswerButton>
 	<p>{props.question.answers[0].user} <span>&#183;</span> {moment(props.question.answers[0].created_at).startOf('hour').fromNow()}</p>
